+++ keyboards/1upkeyboards/1up60hte/1up60hte.c
@@ -16,19 +16,3 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 #include "1up60hte.h"
-
-void keyboard_pre_init_kb(void) {
-	// put your keyboard start-up code here
-	// runs once when the firmware starts up
-    setPinOutput(B6);
-    keyboard_pre_init_user();
-}
-
-void led_set_kb(uint8_t usb_led) {
-	if (IS_LED_ON(usb_led, USB_LED_CAPS_LOCK)) {
-		writePinLow(B6);
-	} else {
-		writePinHigh(B6);
-	}
-    led_set_user(usb_led);
-}
