+++ keyboards/cutie_club/wraith/wraith.c
@@ -15,29 +15,8 @@
  */
 #include "wraith.h"
 
-// Optional override functions below.
-// You can leave any or all of these undefined.
-// These are only required if you want to perform custom actions.
-
 void matrix_init_kb(void) {
-  setPinOutput(B3);
-  setPinOutput(B0);
-  matrix_init_user();
-}
+    setPinOutput(B0);
 
-void matrix_scan_kb(void) {
-  matrix_scan_user();
-}
-
-bool process_record_kb(uint16_t keycode, keyrecord_t *record) {
-  return process_record_user(keycode, record);
-}
-
-void led_set_kb(uint8_t usb_led) {
-  if (IS_LED_ON(usb_led, USB_LED_CAPS_LOCK)) {
-        writePinLow(B3);
-    } else {
-        writePinHigh(B3);
-    }
-  led_set_user(usb_led);
+    matrix_init_user();
 }
