+++ keyboards/omnikeyish/config.h
@@ -24,9 +24,9 @@
 #endif 
 #define MATRIX_COL_PINS { F0, F1, F2, F3, F4, F5, F6, F7, C7, C6, C5, C4, C3, C2, C1, C0, B0, B1, B2, B3, B4, B5, B6 }
 
-#define NUMLOCKLEDPIN E0
-#define CAPSLOCKLEDPIN E1
-#define SCROLLLOCKLEDPIN B7
+#define LED_NUM_LOCK_PIN E0
+#define LED_CAPS_LOCK_PIN E1
+#define LED_SCROLL_LOCK_PIN B7
 
 /* COL2ROW or ROW2COL */
 #define DIODE_DIRECTION ROW2COL
