+++ keyboards/omnikeyish/omnikeyish.c
@@ -1,11 +1,6 @@
 #include "omnikeyish.h"
 
 void keyboard_pre_init_user(void) {
-  /* Configure LED driving pins as output pins */
-  setPinOutput(NUMLOCKLEDPIN);
-  setPinOutput(CAPSLOCKLEDPIN);
-  setPinOutput(SCROLLLOCKLEDPIN);
-
   dynamic_macro_init();
 }
 
@@ -32,24 +27,4 @@ bool process_record_user(uint16_t keycode, keyrecord_t *record) {
   }
 
   return true;
-}
-
-void led_set_user(uint8_t usb_led) {
-  if (IS_LED_ON(usb_led, USB_LED_NUM_LOCK)) {
-    writePinHigh(NUMLOCKLEDPIN);
-  } else {
-    writePinLow(NUMLOCKLEDPIN);
-  }
-
-  if (IS_LED_ON(usb_led, USB_LED_CAPS_LOCK)) {
-    writePinHigh(CAPSLOCKLEDPIN);
-  } else {
-    writePinLow(CAPSLOCKLEDPIN);
-  }
-
-  if (IS_LED_ON(usb_led, USB_LED_SCROLL_LOCK)) {
-    writePinHigh(SCROLLLOCKLEDPIN);
-  } else {
-    writePinLow(SCROLLLOCKLEDPIN);
-  }
 }
