+++ keyboards/tkc/tkc1800/config.h
@@ -43,6 +43,10 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 /* COL2ROW or ROW2COL */
 #define DIODE_DIRECTION COL2ROW
 
+#define LED_NUM_LOCK_PIN D2
+#define LED_CAPS_LOCK_PIN D3
+#define LED_SCROLL_LOCK_PIN D4
+
 #define BACKLIGHT_PIN B6
 #define BACKLIGHT_BREATHING
 #define BACKLIGHT_LEVELS 3
