+++ keyboards/xd84/config.h
@@ -47,6 +47,9 @@
 /* COL2ROW, ROW2COL */
 //#define DIODE_DIRECTION COL2ROW
 
+#define LED_CAPS_LOCK_PIN B6
+#define LED_PIN_ON_STATE 0
+
 #define BACKLIGHT_PIN B5
 #define BACKLIGHT_LEVELS 10
 // #define BACKLIGHT_BREATHING
