+++ keyboards/xd84/xd84.c
@@ -14,19 +14,3 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include "xd84.h"
-
-void keyboard_pre_init_kb(void) {
-  setPinOutput(B6);
-
-  keyboard_pre_init_user();
-}
-
-void led_set_kb(uint8_t usb_led) {
-  if (IS_LED_ON(usb_led, USB_LED_CAPS_LOCK)) {
-    writePinLow(B6);
-  } else {
-    writePinHigh(B6);
-  }
-
-  led_set_user(usb_led);
-}
