+++ keyboards/xd84pro/config.h
@@ -35,6 +35,9 @@
 
 #define DIODE_DIRECTION COL2ROW
 
+#define LED_CAPS_LOCK_PIN B2
+#define LED_PIN_ON_STATE 0
+
 #define RGB_DI_PIN F6
 #ifdef RGB_DI_PIN
     #define RGBLED_NUM 12
