+++ keyboards/xd84pro/xd84pro.c
@@ -14,17 +14,3 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include "xd84pro.h"
-
-void keyboard_pre_init_kb(void) {
-    led_init_ports();
-    keyboard_pre_init_user();
-}
-void led_init_ports(void) {
-    setPinOutput(B2);
-}
-bool led_update_kb(led_t led_state) {
-    if (led_update_user(led_state)) {
-        writePin(B2, !led_state.caps_lock);
-    }
-    return false;
-}
