+++ keyboards/xd96/config.h
@@ -47,6 +47,10 @@
 /* COL2ROW, ROW2COL */
 //#define DIODE_DIRECTION COL2ROW
 
+#define LED_NUM_LOCK_PIN C6
+#define LED_CAPS_LOCK_PIN B6
+#define LED_PIN_ON_STATE 0
+
 #define BACKLIGHT_PIN B5
 #define BACKLIGHT_LEVELS 10
 // #define BACKLIGHT_BREATHING
