+++ keyboards/xd96/xd96.c
@@ -14,29 +14,3 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 #include "xd96.h"
-
-#define CAPS_PIN B6
-#define NUMLOCK_PIN C6
-
-void keyboard_pre_init_kb(void) {
-  setPinOutput(CAPS_PIN);
-  setPinOutput(NUMLOCK_PIN);
-
-  keyboard_pre_init_user();
-}
-
-void led_set_kb(uint8_t usb_led) {
-  if (IS_LED_ON(usb_led, USB_LED_CAPS_LOCK)) {
-    writePinLow(CAPS_PIN);
-  } else {
-    writePinHigh(CAPS_PIN);
-  }
-  
-  if (IS_LED_ON(usb_led, USB_LED_NUM_LOCK)) {
-    writePinLow(NUMLOCK_PIN);
-  } else {
-    writePinHigh(NUMLOCK_PIN);
-  }
-
-  led_set_user(usb_led);
-}
