+++ keyboards/z150_bh/config.h
@@ -22,6 +22,10 @@
 /* COL2ROW or ROW2COL */
 #define DIODE_DIRECTION COL2ROW
 
+#define LED_NUM_LOCK_PIN E7
+#define LED_CAPS_LOCK_PIN B0
+#define LED_SCROLL_LOCK_PIN E6
+
 /* Set 0 if debouncing isn't needed */
 #define DEBOUNCE 5
